requirements.txt
# ExoIntel Platform Dependencies

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database Connectivity
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Machine Learning & AI
scikit-learn>=1.2.0
joblib>=1.2.0
shap>=0.41.0

# Visualization & Frontend
streamlit>=1.22.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0